Removed parallel downloader and did minor refactoring/restructuring
